// src/types/chair.ts
export interface Chair {
  id: number;

  serial_number: string;

  device_id: string;

  model_name: string;

  store_name: string;

  city: string;

  owner_name: string;

  installed_days: number;

  daily_revenue: number;

  monthly_revenue: number;

  status: string;

  performance_score: number;
}